pyproject.toml
[tool.black]
line-length = 120

[tool.ruff]
line-length = 140
ignore = ["E402"]

[tool.poetry]
name = "aoc"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
markdownify = "^0.14.1"
more-itertools = "^10.5.0"
numpy = "^2.2.0"
python-dotenv = "^1.2.1"


[tool.poetry.group.dev.dependencies]
debugpy = "^1.8.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
